use jsz instead of struct direct in final state test
